[0.2] refactor components, start add new style for StartPage
