--- src/pages/start/StartPage.css
@@ -1,129 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Raleway&display=swap");
-
-* {
-  -webkit-user-select:none;
-  -khtml-user-select:none;
-  -moz-user-select:none;
-  -ms-user-select:none;
-  user-select:none;
-}
-
-body {
-  margin: 40px 80px;
-  padding: 0;
-  font-family: "Raleway", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-button {
-  min-width: 422px;
-  background: transparent;
-  border: 1px solid #777;
-  border-radius: 4px;
-}
-
-button:hover {
-  background: mediumseagreen;
-  color: white;
-  border: 1px solid mediumseagreen;
-}
-
-button:disabled {
-  border: 1px solid #999999;
-  background-color: #cccccc;
-  color: #666666;
-}
-
-select {
-  width: 422px;
-  overflow: hidden;
-  border-radius: 3px;
-}
-
-select:hover {
-  -webkit-box-shadow: 0px 0px 20px 0px rgba(60, 179, 113, 1);
-  -moz-box-shadow: 0px 0px 20px 0px rgba(60, 179, 113, 1);
-  box-shadow: 0px 0px 20px 0px rgb(60, 179, 113);
-}
-
-select:focus {
-  outline: none !important;
-  border: 1px solid mediumseagreen;
-  box-shadow: 0 0 10px mediumseagreen;
-}
-
-option {
-  padding: 10px;
-  font-size: 14px;
-}
-
-select:focus option:checked {
-  background: linear-gradient(0deg, mediumseagreen 0%, mediumseagreen 100%);
-}
-
-textarea {
-  min-width: 400px;
-  cursor: pointer;
-  border-radius: 3px;
-
-  padding: 10px;
-}
-
-textarea:focus {
-  outline: none !important;
-  border: 1px solid mediumseagreen;
-  box-shadow: 0 0 10px mediumseagreen;
-}
-
-textarea:hover {
-  -webkit-box-shadow: 0px 0px 20px 0px rgba(60, 179, 113, 1);
-  -moz-box-shadow: 0px 0px 20px 0px rgba(60, 179, 113, 1);
-  box-shadow: 0px 0px 20px 0px rgba(60, 179, 113, 1);
-}
-
-.about {
-  font-size: 13px;
-  margin-bottom: 70px;
-}
-
-/* Scrollbar styles */
-::-webkit-scrollbar {
-  width: 8px;
-  height: 8px;
-}
-
-::-webkit-scrollbar-track {
-  background: transparent;
-  border-radius: 10px;
-}
-
-::-webkit-scrollbar-thumb {
-  border-radius: 10px;
-  background: #ccc;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background: #999;
-}
-
-@media screen and (max-width: 1024px) {
-  body {
-    margin-left: 20px;
-    margin-right: 20px;
-  }
-
-  button {
-    min-width: 80%;
-    padding: 20px 40px;
-  }
-
-  select {
-    min-width: 80%;
-    width: auto;
-  }
-
-  textarea {
-    min-width: 80%;
-  }
-}
