+++ src/pages/start/StartPage.js
@@ -1,133 +1,21 @@
-import React, { useEffect, useState, useRef } from "react";
-import TestPage from "../test/TestPage";
-import Header from './Header';
-import Select from './Select';
-import Switch from "../../components/Switch/Switch";
+import React from 'react';
 
-import tests from '../../services/tests';
-import { downloadFile } from "../../utils/utils";
-import "./StartPage.css";
-import Alert from '../../components/Alert/Alert';
+import styles from './startpage.module.css'
+import TestSelect from '../../components/TestSelect/TestSelect';
+import Prolog from '../../components/Prolog/Prolog';
+import {cn} from '../../utils/utils';
 
 const StartPage = () => {
-  const [select, setSelect] = useState(null);
-  const [testOn, setTestOn] = useState(false);
-  const [buttonOff, setButtonOff] = useState(true);
-  const [userTest, setUserTest] = useState("");
-  const [withMicro, setWithMicro] = useState(false)
-  const [error, setError] = useState(false)
-
-  const recorder = useRef(null)
-  const voice = useRef([])
-
-  useEffect(() => {
-    if (withMicro && !recorder.current) {
-      navigator.mediaDevices.getUserMedia({audio: true})
-        .then(stream => {
-          recorder.current = new MediaRecorder(stream)
-          recorder.current.addEventListener('dataavailable', (e) => {
-            voice.current.push(e.data)
-          })
-          recorder.current.addEventListener('stop', () => {
-            const voiceBlob = new Blob(voice.current, {
-              type: 'audio/mp3'
-            })
-            downloadFile(voiceBlob)
-          })
-        })
-        .catch(e => {
-          setError(true)
-        })
-    }
-  }, [withMicro])
-
-  const startRecorder = () => {
-    if (withMicro && recorder.current) {
-      recorder.current.start()
-    }
-  }
-
-  const stopRecorder = () => {
-    if (withMicro && recorder.current) {
-      recorder.current.stop()
-    }
-  }
-
-  const displayTest = () => {
-    setTestOn(true);
-  };
-
-  const handleSelect = (id) => {
-    setSelect(id);
-  };
-
-  const handleTextarea = (event) => {
-    setUserTest(event.target.value);
-  };
-
-  const buttonTitle = () => {
-    if (select === null) return "Выбирете тест";
-    return buttonOff ? "Напишите хотя бы 3 вопроса для своего теста" : null;
-  };
-
-  const handleSwitch = () => {
-    withMicro ? setWithMicro(false) : setWithMicro(true)
-  }
-
-  const handleError = () => {
-    error ? setError(false) : setError(true)
-  }
-
-  useEffect(() => {
-    if (select !== null) {
-      setButtonOff(false);
-    }
-    tests.UserTest = userTest.split("\n");
-
-    if (select === "UserTest" && tests.UserTest.length < 3) {
-      setButtonOff(true);
-    }
-  }, [select, userTest]);
   return (
-    <>
-      {!testOn && (
-        <div>
-          <Header />
-          <form>
-            <Select
-              select={select}
-              userTest={userTest}
-              handleSelect={handleSelect}
-              handleTextarea={handleTextarea}
-            />
-            {error && (<Alert
-              text={'Микрофон не подключен'}
-              handleError={handleError}
-            />)}
-            <Switch
-              handleSwitch={handleSwitch}
-            />
-            <p>
-              <button
-                type="submit"
-                title={buttonTitle()}
-                disabled={buttonOff}
-                onClick={() => displayTest()}
-              >
-                Начать интервью
-              </button>
-            </p>
-          </form>
-        </div>
-      )}
-      {testOn && (
-        <TestPage
-          initQuestions={tests[select]}
-          startRecorder={startRecorder}
-          stopRecorder={stopRecorder} />
-      )}
-    </>
-  );
-};
+    <div className={styles.container}>
+      <div className={cn(styles.row, styles.left)}>
+        <Prolog />
+      </div>
+      <div className={cn(styles.row, styles.right)}>
+        <TestSelect />
+      </div>
+    </div>
+  )
+}
 
-export default StartPage;
+export default StartPage
